Removed tenant prefix from commands
