+++ docs/manual/docs/cmd/spo/cdn-origin-remove.md
@@ -1,11 +1,11 @@
-# spo tenant cdn get
+# spo cdn origin remove
 
-View current status of the specified Office 365 CDN
+Removes CDN origin for the current SharePoint Online tenant
 
 ## Usage
 
 ```sh
-spo tenant cdn get [options]
+spo cdn origin remove [options]
 ```
 
 ## Options
@@ -14,6 +14,8 @@ Option|Description
 ------|-----------
 `--help`|output usage information
 `-t, --type [type]`|Type of CDN to manage. `Public|Private`. Default `Public`
+`-o, --origin <origin>`|Origin to remove from the current CDN configuration
+`--confirm`|Don't prompt for confirming removal of a tenant property
 `--verbose`|Runs command with verbose logging
 
 !!! important
@@ -21,7 +23,7 @@ Option|Description
 
 ## Remarks
 
-To view the status of an Office 365 CDN, you have to first connect to a tenant admin site using the
+To remove an origin from an Office 365 CDN, you have to first connect to a tenant admin site using the
 [spo connect](connect.md) command, eg. `spo connect https://contoso-admin.sharepoint.com`.
 If you are connected to a different site and will try to manage tenant properties,
 you will get an error.
@@ -32,16 +34,10 @@ the Public (default) or Private CDN. If you don't use the option, the command wi
 ## Examples
 
 ```sh
-spo tenant cdn get
+spo cdn origin remove -t Public -o */CDN
 ```
 
-shows if the Public CDN is currently enabled or not
-
-```sh
-spo tenant cdn get -t Private
-```
-
-shows if the Private CDN is currently enabled or not
+removes */CDN from the list of origins of the Public CDN
 
 ## More information
