+++ docs/manual/docs/cmd/spo/cdn-origin-set.md
@@ -1,11 +1,11 @@
-# spo tenant cdn origin remove
+# spo cdn origin set
 
-Removes CDN origin for the current SharePoint Online tenant
+Adds CDN origin to the current SharePoint Online tenant
 
 ## Usage
 
 ```sh
-spo tenant cdn origin remove [options]
+spo cdn origin set [options]
 ```
 
 ## Options
@@ -14,8 +14,7 @@ Option|Description
 ------|-----------
 `--help`|output usage information
 `-t, --type [type]`|Type of CDN to manage. `Public|Private`. Default `Public`
-`-o, --origin <origin>`|Origin to remove from the current CDN configuration
-`--confirm`|Don't prompt for confirming removal of a tenant property
+`-o, --origin <origin>`|Origin to add to the current CDN configuration
 `--verbose`|Runs command with verbose logging
 
 !!! important
@@ -23,7 +22,7 @@ Option|Description
 
 ## Remarks
 
-To remove an origin from an Office 365 CDN, you have to first connect to a tenant admin site using the
+To add origins to an Office 365 CDN, you have to first connect to a tenant admin site using the
 [spo connect](connect.md) command, eg. `spo connect https://contoso-admin.sharepoint.com`.
 If you are connected to a different site and will try to manage tenant properties,
 you will get an error.
@@ -34,10 +33,10 @@ the Public (default) or Private CDN. If you don't use the option, the command wi
 ## Examples
 
 ```sh
-spo tenant cdn origin remove -t Public -o */CDN
+spo cdn origin set -t Public -o */CDN
 ```
 
-removes */CDN from the list of origins of the Public CDN
+adds */CDN to the list of origins of the Public CDN
 
 ## More information
