+++ docs/manual/docs/cmd/spo/cdn-set.md
@@ -1,11 +1,11 @@
-# spo tenant cdn policy list
+# spo cdn set
 
-Lists CDN policies settings for the current SharePoint Online tenant
+Enable or disable the specified Office 365 CDN
 
 ## Usage
 
 ```sh
-spo tenant cdn policy list [options]
+spo cdn set [options]
 ```
 
 ## Options
@@ -13,6 +13,7 @@ spo tenant cdn policy list [options]
 Option|Description
 ------|-----------
 `--help`|output usage information
+`-e, --enabled <enabled>`|Set to true to enable CDN or to false to disable it. Valid values are true|false
 `-t, --type [type]`|Type of CDN to manage. `Public|Private`. Default `Public`
 `--verbose`|Runs command with verbose logging
 
@@ -21,7 +22,7 @@ Option|Description
 
 ## Remarks
 
-To list the policies of an Office 365 CDN, you have to first connect to a tenant admin site using the
+To enable or disable an Office 365 CDN, you have to first connect to a tenant admin site using the
 [spo connect](connect.md) command, eg. `spo connect https://contoso-admin.sharepoint.com`.
 If you are connected to a different site and will try to manage tenant properties,
 you will get an error.
@@ -29,19 +30,23 @@ you will get an error.
 Using the `-t, --type` option you can choose whether you want to manage the settings of
 the Public (default) or Private CDN. If you don't use the option, the command will use the Public CDN.
 
+Using the `-e, --enabled` option you can specify whether the given CDN type should be
+enabled or disabled. Use true to enable the specified CDN and false to
+disable it.
+
 ## Examples
 
 ```sh
-spo tenant cdn policy list
+spo cdn set -t Public -e true
 ```
 
-shows the list of policies configured for the Public CDN
+enables the Office 365 Public CDN on the current tenant
 
 ```sh
-spo tenant cdn policy list -t Private
+spo cdn set -t Public -e false
 ```
 
-shows the list of policies configured for the Private CDN
+disables the Office 365 Public CDN on the current tenant
 
 ## More information
