+++ src/o365/spo/commands/cdn-get.spec.ts
@@ -1,15 +1,15 @@
 import commands from '../commands';
 import Command, { CommandHelp, CommandValidate, CommandOption } from '../../../Command';
+import config from '../../../config';
 import * as sinon from 'sinon';
 import appInsights from '../../../appInsights';
 import auth, { Site } from '../SpoAuth';
-const tenantCdnPolicyListCommand: Command = require('./tenant-cdn-policy-list');
+const command: Command = require('./cdn-get');
 import * as assert from 'assert';
 import * as request from 'request-promise-native';
 import Utils from '../../../Utils';
-import config from '../../../config';
 
-describe(commands.TENANT_CDN_POLICY_LIST, () => {
+describe(commands.CDN_GET, () => {
   let vorpal: Vorpal;
   let log: string[];
   let cmdInstance: any;
@@ -36,10 +36,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
   });
 
   afterEach(() => {
-    Utils.restore([
-      vorpal.find,
-      request.post
-    ]);
+    Utils.restore(vorpal.find);
   });
 
   after(() => {
@@ -50,15 +47,15 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
   });
 
   it('has correct name', () => {
-    assert.equal(tenantCdnPolicyListCommand.name.startsWith(commands.TENANT_CDN_POLICY_LIST), true);
+    assert.equal(command.name.startsWith(commands.CDN_GET), true);
   });
 
   it('has a description', () => {
-    assert.notEqual(tenantCdnPolicyListCommand.description, null);
+    assert.notEqual(command.description, null);
   });
 
   it('calls telemetry', (done) => {
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       try {
         assert(trackEvent.called);
@@ -71,10 +68,10 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
   });
 
   it('logs correct telemetry event', (done) => {
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       try {
-        assert.equal(telemetry.name, commands.TENANT_CDN_POLICY_LIST);
+        assert.equal(telemetry.name, commands.CDN_GET);
         done();
       }
       catch (e) {
@@ -86,7 +83,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
   it('aborts when not connected to a SharePoint site', (done) => {
     auth.site = new Site();
     auth.site.connected = false;
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       let returnsCorrectValue: boolean = false;
       log.forEach(l => {
@@ -108,7 +105,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso.sharepoint.com';
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       let returnsCorrectValue: boolean = false;
       log.forEach(l => {
@@ -126,7 +123,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     });
   });
 
-  it('retrieves the policies for the public CDN when type set to Public', (done) => {
+  it('retrieves the settings of the public CDN when type set to Public', (done) => {
     sinon.stub(request, 'post').callsFake((opts) => {
       if (opts.url.indexOf('/_api/contextinfo') > -1) {
         if (opts.headers.authorization &&
@@ -142,8 +139,8 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.headers['X-RequestDigest'] === 'abc' &&
-          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnPolicies" Id="7" ObjectPathId="3"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="3" Name="abc" /></ObjectPaths></Request>`) {
-          return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7025.1207", "ErrorInfo": null, "TraceCorrelationId": "0c93299e-7040-4000-83c8-74ba9192542c" }, 7, ["IncludeFileExtensions;CSS,EOT,GIF,ICO,JPEG,JPG,JS,MAP,PNG,SVG,TTF,WOFF,JSON", "ExcludeRestrictedSiteClassifications; ", "ExcludeIfNoScriptDisabled;False"]]));
+          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
+          return Promise.resolve(JSON.stringify([{"SchemaVersion":"15.0.0.0","LibraryVersion":"16.0.7025.1207","ErrorInfo":null,"TraceCorrelationId":"3d92299e-e019-4000-c866-de7d45aa9628"},12,true]));
         }
       }
 
@@ -154,23 +151,20 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true, type: 'Public' } }, () => {
-      let correctLogStatements = 0;
+      let correctLogStatement = false;
       log.forEach(l => {
         if (!l || typeof l !== 'string') {
           return;
         }
 
-        if (l.indexOf('Configured policies:') > -1 ||
-          l.indexOf('IncludeFileExtensions') > -1 ||
-          l.indexOf('ExcludeRestrictedSiteClassifications') > -1 ||
-          l.indexOf('ExcludeIfNoScriptDisabled') > -1) {
-          correctLogStatements++;
+        if (l.indexOf('Public CDN at') > -1 && l.indexOf('enabled') > -1) {
+          correctLogStatement = true;
         }
       });
       try {
-        assert.equal(correctLogStatements, 5);
+        assert(correctLogStatement);
         done();
       }
       catch (e) {
@@ -182,7 +176,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     });
   });
 
-  it('retrieves the policies for the private CDN when type set to Private', (done) => {
+  it('retrieves the settings of the private CDN when type set to Private', (done) => {
     sinon.stub(request, 'post').callsFake((opts) => {
       if (opts.url.indexOf('/_api/contextinfo') > -1) {
         if (opts.headers.authorization &&
@@ -198,8 +192,8 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.headers['X-RequestDigest'] === 'abc' &&
-          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnPolicies" Id="7" ObjectPathId="3"><Parameters><Parameter Type="Enum">1</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="3" Name="abc" /></ObjectPaths></Request>`) {
-          return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7025.1207", "ErrorInfo": null, "TraceCorrelationId": "0c93299e-7040-4000-83c8-74ba9192542c" }, 7, ["IncludeFileExtensions;CSS,EOT,GIF,ICO,JPEG,JPG,JS,MAP,PNG,SVG,TTF,WOFF,JSON", "ExcludeRestrictedSiteClassifications; ", "ExcludeIfNoScriptDisabled;False"]]));
+          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">1</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
+          return Promise.resolve(JSON.stringify([{"SchemaVersion":"15.0.0.0","LibraryVersion":"16.0.7025.1207","ErrorInfo":null,"TraceCorrelationId":"3d92299e-e019-4000-c866-de7d45aa9628"},12,false]));
         }
       }
 
@@ -210,23 +204,20 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
-    cmdInstance.action({ options: { verbose: true, type: 'Private' } }, () => {
-      let correctLogStatements = 0;
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: false, type: 'Private' } }, () => {
+      let correctLogStatement = false;
       log.forEach(l => {
         if (!l || typeof l !== 'string') {
           return;
         }
 
-        if (l.indexOf('Configured policies:') > -1 ||
-          l.indexOf('IncludeFileExtensions') > -1 ||
-          l.indexOf('ExcludeRestrictedSiteClassifications') > -1 ||
-          l.indexOf('ExcludeIfNoScriptDisabled') > -1) {
-          correctLogStatements++;
+        if (l.indexOf('Private CDN at') > -1 && l.indexOf('disabled') > -1) {
+          correctLogStatement = true;
         }
       });
       try {
-        assert.equal(correctLogStatements, 5);
+        assert(correctLogStatement);
         done();
       }
       catch (e) {
@@ -238,7 +229,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     });
   });
 
-  it('retrieves the policies for the public CDN when no type set', (done) => {
+  it('retrieves the settings of the public CDN when no type set', (done) => {
     sinon.stub(request, 'post').callsFake((opts) => {
       if (opts.url.indexOf('/_api/contextinfo') > -1) {
         if (opts.headers.authorization &&
@@ -254,8 +245,8 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.headers['X-RequestDigest'] === 'abc' &&
-          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnPolicies" Id="7" ObjectPathId="3"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="3" Name="abc" /></ObjectPaths></Request>`) {
-          return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7025.1207", "ErrorInfo": null, "TraceCorrelationId": "0c93299e-7040-4000-83c8-74ba9192542c" }, 7, ["IncludeFileExtensions;CSS,EOT,GIF,ICO,JPEG,JPG,JS,MAP,PNG,SVG,TTF,WOFF,JSON", "ExcludeRestrictedSiteClassifications; ", "ExcludeIfNoScriptDisabled;False"]]));
+          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
+          return Promise.resolve(JSON.stringify([{"SchemaVersion":"15.0.0.0","LibraryVersion":"16.0.7025.1207","ErrorInfo":null,"TraceCorrelationId":"3d92299e-e019-4000-c866-de7d45aa9628"},12,true]));
         }
       }
 
@@ -266,23 +257,20 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
-    cmdInstance.action({ options: { verbose: false } }, () => {
-      let correctLogStatements = 0;
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: true } }, () => {
+      let correctLogStatement = false;
       log.forEach(l => {
         if (!l || typeof l !== 'string') {
           return;
         }
 
-        if (l.indexOf('Configured policies:') > -1 ||
-          l.indexOf('IncludeFileExtensions') > -1 ||
-          l.indexOf('ExcludeRestrictedSiteClassifications') > -1 ||
-          l.indexOf('ExcludeIfNoScriptDisabled') > -1) {
-          correctLogStatements++;
+        if (l.indexOf('Public CDN at') > -1 && l.indexOf('enabled') > -1) {
+          correctLogStatement = true;
         }
       });
       try {
-        assert.equal(correctLogStatements, 4);
+        assert(correctLogStatement);
         done();
       }
       catch (e) {
@@ -294,7 +282,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     });
   });
 
-  it('correctly handles an error when retrieving tenant CDN policies', (done) => {
+  it('correctly handles an error when getting tenant CDN settings', (done) => {
     Utils.restore(request.post);
     sinon.stub(request, 'post').callsFake((opts) => {
       if (opts.url.indexOf('/_api/contextinfo') > -1) {
@@ -311,7 +299,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnPolicies" Id="7" ObjectPathId="3"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="3" Name="abc" /></ObjectPaths></Request>`) {
+          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
             return Promise.resolve(JSON.stringify([
               {
                 "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": {
@@ -330,8 +318,8 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
-    cmdInstance.action({ options: { verbose: false } }, () => {
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: true } }, () => {
       let genericErrorHandled = false;
       log.forEach(l => {
         if (l && typeof l === 'string' && l.indexOf('An error has occurred') > -1) {
@@ -353,7 +341,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
   });
 
   it('supports verbose mode', () => {
-    const options = (tenantCdnPolicyListCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     let containsVerboseOption = false;
     options.forEach(o => {
       if (o.option === '--verbose') {
@@ -364,7 +352,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
   });
 
   it('supports specifying CDN type', () => {
-    const options = (tenantCdnPolicyListCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     let containsTypeOption = false;
     options.forEach(o => {
       if (o.option.indexOf('[type]') > -1) {
@@ -376,29 +364,29 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
 
   it('doesn\'t fail if the parent doesn\'t define options', () => {
     sinon.stub(Command.prototype, 'options').callsFake(() => { return undefined; });
-    const options = (tenantCdnPolicyListCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     Utils.restore(Command.prototype.options);
     assert(options.length > 0);
   });
 
   it('accepts Public SharePoint Online CDN type', () => {
-    const actual = (tenantCdnPolicyListCommand.validate() as CommandValidate)({ options: { type: 'Public' } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: 'Public' } });
     assert(actual);
   });
 
   it('accepts Private SharePoint Online CDN type', () => {
-    const actual = (tenantCdnPolicyListCommand.validate() as CommandValidate)({ options: { type: 'Private' } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: 'Private' } });
     assert(actual);
   });
 
   it('rejects invalid SharePoint Online CDN type', () => {
     const type = 'foo';
-    const actual = (tenantCdnPolicyListCommand.validate() as CommandValidate)({ options: { type: type } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: type } });
     assert.equal(actual, `${type} is not a valid CDN type. Allowed values are Public|Private`);
   });
 
   it('doesn\'t fail validation if the optional type option not specified', () => {
-    const actual = (tenantCdnPolicyListCommand.validate() as CommandValidate)({ options: {} });
+    const actual = (command.validate() as CommandValidate)({ options: {} });
     assert(actual);
   });
 
@@ -409,8 +397,8 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
       helpInformation: () => { }
     };
     const find = sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnPolicyListCommand.help() as CommandHelp)({}, helpLog);
-    assert(find.calledWith(commands.TENANT_CDN_POLICY_LIST));
+    (command.help() as CommandHelp)({}, helpLog);
+    assert(find.calledWith(commands.CDN_GET));
   });
 
   it('has help with examples', () => {
@@ -420,7 +408,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
       helpInformation: () => { }
     };
     sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnPolicyListCommand.help() as CommandHelp)({}, log);
+    (command.help() as CommandHelp)({}, log);
     let containsExamples: boolean = false;
     _log.forEach(l => {
       if (l && l.indexOf('Examples:') > -1) {
@@ -437,7 +425,7 @@ describe(commands.TENANT_CDN_POLICY_LIST, () => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
-    cmdInstance.action = tenantCdnPolicyListCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       let containsError = false;
       log.forEach(l => {
