+++ src/o365/spo/commands/cdn-origin-list.spec.ts
@@ -1,15 +1,15 @@
 import commands from '../commands';
 import Command, { CommandHelp, CommandValidate, CommandOption } from '../../../Command';
-import config from '../../../config';
 import * as sinon from 'sinon';
 import appInsights from '../../../appInsights';
 import auth, { Site } from '../SpoAuth';
-const tenantCdnGetCommand: Command = require('./tenant-cdn-get');
+const command: Command = require('./cdn-origin-list');
 import * as assert from 'assert';
 import * as request from 'request-promise-native';
 import Utils from '../../../Utils';
+import config from '../../../config';
 
-describe(commands.TENANT_CDN_GET, () => {
+describe(commands.CDN_ORIGIN_LIST, () => {
   let vorpal: Vorpal;
   let log: string[];
   let cmdInstance: any;
@@ -36,7 +36,10 @@ describe(commands.TENANT_CDN_GET, () => {
   });
 
   afterEach(() => {
-    Utils.restore(vorpal.find);
+    Utils.restore([
+      vorpal.find,
+      request.get
+    ]);
   });
 
   after(() => {
@@ -47,15 +50,15 @@ describe(commands.TENANT_CDN_GET, () => {
   });
 
   it('has correct name', () => {
-    assert.equal(tenantCdnGetCommand.name.startsWith(commands.TENANT_CDN_GET), true);
+    assert.equal(command.name.startsWith(commands.CDN_ORIGIN_LIST), true);
   });
 
   it('has a description', () => {
-    assert.notEqual(tenantCdnGetCommand.description, null);
+    assert.notEqual(command.description, null);
   });
 
   it('calls telemetry', (done) => {
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       try {
         assert(trackEvent.called);
@@ -68,10 +71,10 @@ describe(commands.TENANT_CDN_GET, () => {
   });
 
   it('logs correct telemetry event', (done) => {
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       try {
-        assert.equal(telemetry.name, commands.TENANT_CDN_GET);
+        assert.equal(telemetry.name, commands.CDN_ORIGIN_LIST);
         done();
       }
       catch (e) {
@@ -83,7 +86,7 @@ describe(commands.TENANT_CDN_GET, () => {
   it('aborts when not connected to a SharePoint site', (done) => {
     auth.site = new Site();
     auth.site.connected = false;
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       let returnsCorrectValue: boolean = false;
       log.forEach(l => {
@@ -105,7 +108,7 @@ describe(commands.TENANT_CDN_GET, () => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso.sharepoint.com';
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       let returnsCorrectValue: boolean = false;
       log.forEach(l => {
@@ -139,8 +142,8 @@ describe(commands.TENANT_CDN_GET, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.headers['X-RequestDigest'] === 'abc' &&
-          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
-          return Promise.resolve(JSON.stringify([{"SchemaVersion":"15.0.0.0","LibraryVersion":"16.0.7025.1207","ErrorInfo":null,"TraceCorrelationId":"3d92299e-e019-4000-c866-de7d45aa9628"},12,true]));
+          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnOrigins" Id="22" ObjectPathId="18"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="18" Name="abc" /></ObjectPaths></Request>`) {
+          return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7025.1207", "ErrorInfo": null, "TraceCorrelationId": "8992299e-a003-4000-7686-fda36e26a53c" }, 22, ['/master', '*/cdn']]));
         }
       }
 
@@ -151,20 +154,22 @@ describe(commands.TENANT_CDN_GET, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true, type: 'Public' } }, () => {
-      let correctLogStatement = false;
+      let correctLogStatements = 0;
       log.forEach(l => {
         if (!l || typeof l !== 'string') {
           return;
         }
 
-        if (l.indexOf('Public CDN at') > -1 && l.indexOf('enabled') > -1) {
-          correctLogStatement = true;
+        if (l.indexOf('Configured origins:') > -1 ||
+          l.indexOf('/master') > -1 ||
+          l.indexOf('*/cdn') > -1) {
+          correctLogStatements++;
         }
       });
       try {
-        assert(correctLogStatement);
+        assert.equal(correctLogStatements, 4);
         done();
       }
       catch (e) {
@@ -192,8 +197,8 @@ describe(commands.TENANT_CDN_GET, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.headers['X-RequestDigest'] === 'abc' &&
-          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">1</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
-          return Promise.resolve(JSON.stringify([{"SchemaVersion":"15.0.0.0","LibraryVersion":"16.0.7025.1207","ErrorInfo":null,"TraceCorrelationId":"3d92299e-e019-4000-c866-de7d45aa9628"},12,false]));
+          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnOrigins" Id="22" ObjectPathId="18"><Parameters><Parameter Type="Enum">1</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="18" Name="abc" /></ObjectPaths></Request>`) {
+          return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7025.1207", "ErrorInfo": null, "TraceCorrelationId": "8992299e-a003-4000-7686-fda36e26a53c" }, 22, ['/master']]));
         }
       }
 
@@ -204,20 +209,21 @@ describe(commands.TENANT_CDN_GET, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: false, type: 'Private' } }, () => {
-      let correctLogStatement = false;
+      let correctLogStatements = 0;
       log.forEach(l => {
         if (!l || typeof l !== 'string') {
           return;
         }
 
-        if (l.indexOf('Private CDN at') > -1 && l.indexOf('disabled') > -1) {
-          correctLogStatement = true;
+        if (l.indexOf('Configured origins:') > -1 ||
+          l.indexOf('/master') > -1) {
+          correctLogStatements++;
         }
       });
       try {
-        assert(correctLogStatement);
+        assert.equal(correctLogStatements, 2);
         done();
       }
       catch (e) {
@@ -245,8 +251,8 @@ describe(commands.TENANT_CDN_GET, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.headers['X-RequestDigest'] === 'abc' &&
-          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
-          return Promise.resolve(JSON.stringify([{"SchemaVersion":"15.0.0.0","LibraryVersion":"16.0.7025.1207","ErrorInfo":null,"TraceCorrelationId":"3d92299e-e019-4000-c866-de7d45aa9628"},12,true]));
+          opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnOrigins" Id="22" ObjectPathId="18"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="18" Name="abc" /></ObjectPaths></Request>`) {
+          return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7025.1207", "ErrorInfo": null, "TraceCorrelationId": "8992299e-a003-4000-7686-fda36e26a53c" }, 22, ['/master', '*/cdn']]));
         }
       }
 
@@ -257,20 +263,22 @@ describe(commands.TENANT_CDN_GET, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true } }, () => {
-      let correctLogStatement = false;
+      let correctLogStatements = 0;
       log.forEach(l => {
         if (!l || typeof l !== 'string') {
           return;
         }
 
-        if (l.indexOf('Public CDN at') > -1 && l.indexOf('enabled') > -1) {
-          correctLogStatement = true;
+        if (l.indexOf('Configured origins:') > -1 ||
+          l.indexOf('/master') > -1 ||
+          l.indexOf('*/cdn') > -1) {
+          correctLogStatements++;
         }
       });
       try {
-        assert(correctLogStatement);
+        assert.equal(correctLogStatements, 4);
         done();
       }
       catch (e) {
@@ -282,7 +290,7 @@ describe(commands.TENANT_CDN_GET, () => {
     });
   });
 
-  it('correctly handles an error when getting tenant CDN settings', (done) => {
+  it('correctly handles an error when getting tenant CDN origins', (done) => {
     Utils.restore(request.post);
     sinon.stub(request, 'post').callsFake((opts) => {
       if (opts.url.indexOf('/_api/contextinfo') > -1) {
@@ -299,7 +307,7 @@ describe(commands.TENANT_CDN_GET, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="abc" /></ObjectPaths></Request>`) {
+          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnOrigins" Id="22" ObjectPathId="18"><Parameters><Parameter Type="Enum">0</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="18" Name="abc" /></ObjectPaths></Request>`) {
             return Promise.resolve(JSON.stringify([
               {
                 "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": {
@@ -318,7 +326,7 @@ describe(commands.TENANT_CDN_GET, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true } }, () => {
       let genericErrorHandled = false;
       log.forEach(l => {
@@ -341,7 +349,7 @@ describe(commands.TENANT_CDN_GET, () => {
   });
 
   it('supports verbose mode', () => {
-    const options = (tenantCdnGetCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     let containsVerboseOption = false;
     options.forEach(o => {
       if (o.option === '--verbose') {
@@ -352,7 +360,7 @@ describe(commands.TENANT_CDN_GET, () => {
   });
 
   it('supports specifying CDN type', () => {
-    const options = (tenantCdnGetCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     let containsTypeOption = false;
     options.forEach(o => {
       if (o.option.indexOf('[type]') > -1) {
@@ -364,29 +372,29 @@ describe(commands.TENANT_CDN_GET, () => {
 
   it('doesn\'t fail if the parent doesn\'t define options', () => {
     sinon.stub(Command.prototype, 'options').callsFake(() => { return undefined; });
-    const options = (tenantCdnGetCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     Utils.restore(Command.prototype.options);
     assert(options.length > 0);
   });
 
   it('accepts Public SharePoint Online CDN type', () => {
-    const actual = (tenantCdnGetCommand.validate() as CommandValidate)({ options: { type: 'Public' } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: 'Public' } });
     assert(actual);
   });
 
   it('accepts Private SharePoint Online CDN type', () => {
-    const actual = (tenantCdnGetCommand.validate() as CommandValidate)({ options: { type: 'Private' } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: 'Private' } });
     assert(actual);
   });
 
   it('rejects invalid SharePoint Online CDN type', () => {
     const type = 'foo';
-    const actual = (tenantCdnGetCommand.validate() as CommandValidate)({ options: { type: type } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: type } });
     assert.equal(actual, `${type} is not a valid CDN type. Allowed values are Public|Private`);
   });
 
   it('doesn\'t fail validation if the optional type option not specified', () => {
-    const actual = (tenantCdnGetCommand.validate() as CommandValidate)({ options: {} });
+    const actual = (command.validate() as CommandValidate)({ options: {} });
     assert(actual);
   });
 
@@ -397,8 +405,8 @@ describe(commands.TENANT_CDN_GET, () => {
       helpInformation: () => { }
     };
     const find = sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnGetCommand.help() as CommandHelp)({}, helpLog);
-    assert(find.calledWith(commands.TENANT_CDN_GET));
+    (command.help() as CommandHelp)({}, helpLog);
+    assert(find.calledWith(commands.CDN_ORIGIN_LIST));
   });
 
   it('has help with examples', () => {
@@ -408,7 +416,7 @@ describe(commands.TENANT_CDN_GET, () => {
       helpInformation: () => { }
     };
     sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnGetCommand.help() as CommandHelp)({}, log);
+    (command.help() as CommandHelp)({}, log);
     let containsExamples: boolean = false;
     _log.forEach(l => {
       if (l && l.indexOf('Examples:') > -1) {
@@ -425,7 +433,7 @@ describe(commands.TENANT_CDN_GET, () => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
-    cmdInstance.action = tenantCdnGetCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       let containsError = false;
       log.forEach(l => {
