+++ src/o365/spo/commands/cdn-origin-list.ts
@@ -21,13 +21,13 @@ interface Options extends VerboseOption {
   type: string;
 }
 
-class SpoTenantCdnGetCommand extends SpoCommand {
+class SpoCdnOriginListCommand extends SpoCommand {
   public get name(): string {
-    return commands.TENANT_CDN_GET;
+    return commands.CDN_ORIGIN_LIST;
   }
 
   public get description(): string {
-    return 'View current status of the specified Office 365 CDN';
+    return 'List CDN origins settings for the current SharePoint Online tenant';
   }
 
   public getTelemetryProperties(args: CommandArgs): any {
@@ -44,38 +44,29 @@ class SpoTenantCdnGetCommand extends SpoCommand {
     const cdnTypeString: string = args.options.type || 'Public';
     const cdnType: number = cdnTypeString === 'Private' ? 1 : 0;
 
+    if (this.verbose) {
+      cmd.log(`Retrieving access token for ${auth.service.resource}...`);
+    }
+
     auth
       .ensureAccessToken(auth.service.resource, cmd, this.verbose)
       .then((accessToken: string): Promise<ContextInfo> => {
         if (this.verbose) {
-          cmd.log(`Retrieved access token ${accessToken}. Loading CDN settings for the ${auth.site.url} tenant...`);
-        }
-
-        const requestOptions: any = {
-          url: `${auth.site.url}/_api/contextinfo`,
-          headers: {
-            authorization: `Bearer ${accessToken}`,
-            accept: 'application/json;odata=nometadata'
-          },
-          json: true
-        }
-
-        if (this.verbose) {
-          cmd.log('Executing web request...');
-          cmd.log(requestOptions);
+          cmd.log('Response:');
+          cmd.log(accessToken);
           cmd.log('');
         }
 
-        return request.post(requestOptions);
+        return this.getRequestDigest(cmd, this.verbose);
       })
       .then((res: ContextInfo): Promise<string> => {
         if (this.verbose) {
-          cmd.log('Response:');
+          cmd.log('Response:')
           cmd.log(res);
           cmd.log('');
         }
 
-        cmd.log(`Retrieving status of ${(cdnType === 1 ? 'Private' : 'Public')} CDN...`);
+        cmd.log(`Retrieving configured origins for ${(cdnType === 1 ? 'Private' : 'Public')} CDN...`);
 
         const requestOptions: any = {
           url: `${auth.site.url}/_vti_bin/client.svc/ProcessQuery`,
@@ -83,7 +74,7 @@ class SpoTenantCdnGetCommand extends SpoCommand {
             authorization: `Bearer ${auth.service.accessToken}`,
             'X-RequestDigest': res.FormDigestValue
           },
-          body: `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnEnabled" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">${cdnType}</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="${auth.site.tenantId}" /></ObjectPaths></Request>`
+          body: `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnOrigins" Id="22" ObjectPathId="18"><Parameters><Parameter Type="Enum">${cdnType}</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="18" Name="${auth.site.tenantId}" /></ObjectPaths></Request>`
         };
 
         if (this.verbose) {
@@ -107,8 +98,11 @@ class SpoTenantCdnGetCommand extends SpoCommand {
           cmd.log(vorpal.chalk.red(`Error: ${response.ErrorInfo.ErrorMessage}`));
         }
         else {
-          const result: boolean = json[json.length - 1];
-          cmd.log(`${(cdnType === 0 ? 'Public' : 'Private')} CDN at ${auth.site.url} is ${(result === true ? 'enabled' : 'disabled')}`);
+          const result: string[] = json[json.length - 1];
+          cmd.log('Configured origins:');
+          result.forEach((o: string): void => {
+            cmd.log(o);
+          });
         }
         cb();
       }, (err: any): void => {
@@ -149,14 +143,14 @@ class SpoTenantCdnGetCommand extends SpoCommand {
   public help(): CommandHelp {
     return function (args: CommandArgs, log: (help: string) => void): void {
       const chalk = vorpal.chalk;
-      log(vorpal.find(commands.TENANT_CDN_GET).helpInformation());
+      log(vorpal.find(commands.CDN_ORIGIN_LIST).helpInformation());
       log(
         `  ${chalk.yellow('Important:')} before using this command, connect to a SharePoint Online tenant admin site,
   using the ${chalk.blue(commands.CONNECT)} command.
         
   Remarks:
 
-    To view the status of an Office 365 CDN, you have to first connect to a tenant admin site using the
+    To list origins of a Office 365 CDN, you have to first connect to a tenant admin site using the
     ${chalk.blue(commands.CONNECT)} command, eg. ${chalk.grey(`${config.delimiter} ${commands.CONNECT} https://contoso-admin.sharepoint.com`)}.
     If you are connected to a different site and will try to manage tenant properties,
     you will get an error.
@@ -166,11 +160,11 @@ class SpoTenantCdnGetCommand extends SpoCommand {
 
   Examples:
   
-    ${chalk.grey(config.delimiter)} ${commands.TENANT_CDN_GET}
-      shows if the Public CDN is currently enabled or not
+    ${chalk.grey(config.delimiter)} ${commands.CDN_ORIGIN_LIST}
+      shows the list of origins configured for the Public CDN
 
-    ${chalk.grey(config.delimiter)} ${commands.TENANT_CDN_GET} -t Private
-      shows if the Private CDN is currently enabled or not
+    ${chalk.grey(config.delimiter)} ${commands.CDN_ORIGIN_LIST} -t Private
+      shows the list of origins configured for the Private CDN
 
   More information:
 
@@ -181,4 +175,4 @@ class SpoTenantCdnGetCommand extends SpoCommand {
   }
 }
 
-module.exports = new SpoTenantCdnGetCommand();
+module.exports = new SpoCdnOriginListCommand();
