+++ src/o365/spo/commands/cdn-origin-set.spec.ts
@@ -3,20 +3,19 @@ import Command, { CommandHelp, CommandValidate, CommandOption } from '../../../C
 import * as sinon from 'sinon';
 import appInsights from '../../../appInsights';
 import auth, { Site } from '../SpoAuth';
-const tenantCdnOriginRemoveCommand: Command = require('./tenant-cdn-origin-remove');
+const command: Command = require('./cdn-origin-set');
 import * as assert from 'assert';
 import * as request from 'request-promise-native';
 import config from '../../../config';
 import Utils from '../../../Utils';
 
-describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
+describe(commands.CDN_ORIGIN_SET, () => {
   let vorpal: Vorpal;
   let log: string[];
   let cmdInstance: any;
   let trackEvent: any;
   let telemetry: any;
   let requests: any[];
-  let promptOptions: any;
 
   before(() => {
     sinon.stub(auth, 'ensureAccessToken').callsFake(() => { return Promise.resolve('ABC'); });
@@ -40,8 +39,14 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="RemoveTenantCdnOrigin" Id="33" ObjectPathId="29"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="29" Name="abc" /></ObjectPaths></Request>`) {
-            return Promise.resolve(JSON.stringify([{ "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": null, "TraceCorrelationId": "4456299e-d09e-4000-ae61-ddde716daa27" }, 31, { "IsNull": false }, 33, { "IsNull": false }, 35, { "IsNull": false }]));
+          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
+            return Promise.resolve(JSON.stringify([
+              {
+                "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": null, "TraceCorrelationId": "a05d299e-0036-4000-8546-cfc42dc07fd2"
+              }, 42, [
+                "*\u002fMASTERPAGE", "*\u002fSTYLE LIBRARY", "*\u002fCLIENTSIDEASSETS", "*\u002fCDN (configuration pending)"
+              ]
+            ]));
           }
         }
       }
@@ -56,16 +61,11 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     cmdInstance = {
       log: (msg: string) => {
         log.push(msg);
-      },
-      prompt: (options: any, cb: (result: { continue: boolean }) => void) => {
-        promptOptions = options;
-        cb({ continue: false });
       }
     };
     auth.site = new Site();
     telemetry = null;
     requests = [];
-    promptOptions = undefined;
   });
 
   afterEach(() => {
@@ -81,16 +81,16 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
   });
 
   it('has correct name', () => {
-    assert.equal(tenantCdnOriginRemoveCommand.name.startsWith(commands.TENANT_CDN_ORIGIN_REMOVE), true);
+    assert.equal(command.name.startsWith(commands.CDN_ORIGIN_SET), true);
   });
 
   it('has a description', () => {
-    assert.notEqual(tenantCdnOriginRemoveCommand.description, null);
+    assert.notEqual(command.description, null);
   });
 
   it('calls telemetry', (done) => {
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.action({ options: {}, url: 'https://contoso-admin.sharepoint.com' }, () => {
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       try {
         assert(trackEvent.called);
         done();
@@ -102,10 +102,10 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
   });
 
   it('logs correct telemetry event', (done) => {
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.action({ options: {}, url: 'https://contoso-admin.sharepoint.com' }, () => {
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       try {
-        assert.equal(telemetry.name, commands.TENANT_CDN_ORIGIN_REMOVE);
+        assert.equal(telemetry.name, commands.CDN_ORIGIN_SET);
         done();
       }
       catch (e) {
@@ -117,7 +117,7 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
   it('aborts when not connected to a SharePoint site', (done) => {
     auth.site = new Site();
     auth.site.connected = false;
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       let returnsCorrectValue: boolean = false;
       log.forEach(l => {
@@ -139,7 +139,7 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso.sharepoint.com';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
       let returnsCorrectValue: boolean = false;
       log.forEach(l => {
@@ -157,26 +157,26 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     });
   });
 
-  it('removes existing CDN origin from the public CDN when Public type specified without prompting with confirmation argument', (done) => {
+  it('sets CDN origin on the public CDN when Public type specified', (done) => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.action({ options: { verbose: false, origin: '*/cdn', confirm: true, type: 'Public' } }, () => {
-      let deleteRequestIssued = false;
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: true, origin: '*/cdn', type: 'Public' } }, () => {
+      let setRequestIssued = false;
       requests.forEach(r => {
         if (r.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1 &&
           r.headers.authorization &&
           r.headers.authorization.indexOf('Bearer ') === 0 &&
           r.headers['X-RequestDigest'] &&
-          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="RemoveTenantCdnOrigin" Id="33" ObjectPathId="29"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="29" Name="abc" /></ObjectPaths></Request>`) {
-          deleteRequestIssued = true;
+          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
+          setRequestIssued = true;
         }
       });
 
       try {
-        assert(deleteRequestIssued);
+        assert(setRequestIssued);
         done();
       }
       catch (e) {
@@ -185,26 +185,26 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     });
   });
 
-  it('removes existing CDN origin from the private CDN when Private type specified without prompting with confirmation argument', (done) => {
+  it('sets CDN origin on the private CDN when Private type specified', (done) => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.action({ options: { verbose: false, origin: '*/cdn', confirm: true, type: 'Private' } }, () => {
-      let deleteRequestIssued = false;
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: true, origin: '*/cdn', type: 'Private' } }, () => {
+      let setRequestIssued = false;
       requests.forEach(r => {
         if (r.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1 &&
           r.headers.authorization &&
           r.headers.authorization.indexOf('Bearer ') === 0 &&
           r.headers['X-RequestDigest'] &&
-          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="RemoveTenantCdnOrigin" Id="33" ObjectPathId="29"><Parameters><Parameter Type="Enum">1</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="29" Name="abc" /></ObjectPaths></Request>`) {
-          deleteRequestIssued = true;
+          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">1</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
+          setRequestIssued = true;
         }
       });
 
       try {
-        assert(deleteRequestIssued);
+        assert(setRequestIssued);
         done();
       }
       catch (e) {
@@ -213,26 +213,26 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     });
   });
 
-  it('removes existing CDN origin from the public CDN when no type specified without prompting with confirmation argument', (done) => {
+  it('sets CDN origin on the public CDN when no type specified', (done) => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.action({ options: { verbose: false, origin: '*/cdn', confirm: true } }, () => {
-      let deleteRequestIssued = false;
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: false, origin: '*/cdn' } }, () => {
+      let setRequestIssued = false;
       requests.forEach(r => {
         if (r.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1 &&
           r.headers.authorization &&
           r.headers.authorization.indexOf('Bearer ') === 0 &&
           r.headers['X-RequestDigest'] &&
-          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="RemoveTenantCdnOrigin" Id="33" ObjectPathId="29"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="29" Name="abc" /></ObjectPaths></Request>`) {
-          deleteRequestIssued = true;
+          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
+          setRequestIssued = true;
         }
       });
 
       try {
-        assert(deleteRequestIssued);
+        assert(setRequestIssued);
         done();
       }
       catch (e) {
@@ -241,79 +241,11 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     });
   });
 
-  it('prompts before removing CDN origin when confirmation argument not passed', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn' } }, () => {
-      let promptIssued = false;
-
-      if (promptOptions && promptOptions.type === 'confirm') {
-        promptIssued = true;
-      }
-
-      try {
-        assert(promptIssued);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('aborts removing CDN origin when prompt not confirmed', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.prompt = (options: any, cb: (result: { continue: boolean }) => void) => {
-      cb({ continue: false });
-    };
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn' } }, () => {
-      try {
-        assert(requests.length === 0);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('removes CDN origin when prompt confirmed', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.prompt = (options: any, cb: (result: { continue: boolean }) => void) => {
-      cb({ continue: true });
-    };
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn' } }, () => {
-      let doneResponse = false;
-      log.forEach(l => {
-        if (l &&
-          typeof l === 'string' &&
-          l.indexOf('DONE') > -1) {
-          doneResponse = true;
-        }
-      });
-
-      try {
-        assert(doneResponse);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('correctly handles an error when removing CDN origin', (done) => {
+  it('correctly handles trying to set CDN origin that has already been set', (done) => {
     Utils.restore(request.post);
     sinon.stub(request, 'post').callsFake((opts) => {
+      requests.push(opts);
+
       if (opts.url.indexOf('/_api/contextinfo') > -1) {
         if (opts.headers.authorization &&
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
@@ -328,13 +260,70 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
           opts.headers.authorization.indexOf('Bearer ') === 0 &&
           opts.headers['X-RequestDigest'] &&
           opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="RemoveTenantCdnOrigin" Id="33" ObjectPathId="29"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="29" Name="abc" /></ObjectPaths></Request>`) {
+          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
             return Promise.resolve(JSON.stringify([
               {
                 "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": {
-                  "ErrorMessage": "An error has occurred", "ErrorValue": null, "TraceCorrelationId": "965d299e-a0c6-4000-8546-cc244881a129", "ErrorCode": -1, "ErrorTypeName": "Microsoft.SharePoint.PublicCdn.TenantCdnAdministrationException"
+                  "ErrorMessage": "The library is already registered as a CDN origin.", "ErrorValue": null, "TraceCorrelationId": "965d299e-a0c6-4000-8546-cc244881a129", "ErrorCode": -1, "ErrorTypeName": "Microsoft.SharePoint.PublicCdn.TenantCdnAdministrationException"
                 }, "TraceCorrelationId": "965d299e-a0c6-4000-8546-cc244881a129"
               }
+            ]));
+          }
+        }
+      }
+
+      return Promise.reject('Invalid request');
+    });
+    auth.site = new Site();
+    auth.site.connected = true;
+    auth.site.url = 'https://contoso-admin.sharepoint.com';
+    auth.site.tenantId = 'abc';
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: true, origin: '*/cdn', type: 'Public' } }, () => {
+      let correctErrorReported = false;
+      log.forEach(l => {
+        if (l && typeof l === 'string' &&
+          l.indexOf('The library is already registered as a CDN origin.') > -1) {
+          correctErrorReported = true;
+        }
+      });
+
+      try {
+        assert(correctErrorReported);
+        done();
+      }
+      catch (e) {
+        done(e);
+      }
+    });
+  });
+
+  it('escapes XML in user input', (done) => {
+    Utils.restore(request.post);
+    sinon.stub(request, 'post').callsFake((opts) => {
+      requests.push(opts);
+
+      if (opts.url.indexOf('/_api/contextinfo') > -1) {
+        if (opts.headers.authorization &&
+          opts.headers.authorization.indexOf('Bearer ') === 0 &&
+          opts.headers.accept &&
+          opts.headers.accept.indexOf('application/json') === 0) {
+          return Promise.resolve({ FormDigestValue: 'abc' });
+        }
+      }
+
+      if (opts.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1) {
+        if (opts.headers.authorization &&
+          opts.headers.authorization.indexOf('Bearer ') === 0 &&
+          opts.headers['X-RequestDigest'] &&
+          opts.body) {
+          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">&lt;*/CDN&gt;</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
+            return Promise.resolve(JSON.stringify([
+              {
+                "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": null, "TraceCorrelationId": "a05d299e-0036-4000-8546-cfc42dc07fd2"
+              }, 42, [
+                "*\u002fMASTERPAGE", "*\u002fSTYLE LIBRARY", "*\u002fCLIENTSIDEASSETS", "*\u002fCDN (configuration pending)"
+              ]
             ]));
           }
         }
@@ -347,33 +336,27 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
     auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
-    cmdInstance.prompt = (options: any, cb: (result: { continue: boolean }) => void) => {
-      cb({ continue: true });
-    };
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn' } }, () => {
-      let genericErrorHandled = false;
+    cmdInstance.action = command.action();
+    cmdInstance.action({ options: { verbose: true, origin: '<*/CDN>' } }, () => {
+      let isDone = false;
       log.forEach(l => {
-        if (l && typeof l === 'string' && l.indexOf('An error has occurred') > -1) {
-          genericErrorHandled = true;
+        if (l && typeof l === 'string' && l.indexOf('DONE')) {
+          isDone = true;
         }
       });
 
       try {
-        assert(genericErrorHandled);
+        assert(isDone);
         done();
       }
       catch (e) {
         done(e);
       }
-      finally {
-        Utils.restore(request.post);
-      }
     });
   });
 
   it('supports verbose mode', () => {
-    const options = (tenantCdnOriginRemoveCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     let containsVerboseOption = false;
     options.forEach(o => {
       if (o.option === '--verbose') {
@@ -383,19 +366,8 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     assert(containsVerboseOption);
   });
 
-  it('supports suppressing confirmation prompt', () => {
-    const options = (tenantCdnOriginRemoveCommand.options() as CommandOption[]);
-    let containsConfirmOption = false;
-    options.forEach(o => {
-      if (o.option.indexOf('--confirm') > -1) {
-        containsConfirmOption = true;
-      }
-    });
-    assert(containsConfirmOption);
-  });
-
   it('requires CDN origin name', () => {
-    const options = (tenantCdnOriginRemoveCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     let requiresCdnOriginName = false;
     options.forEach(o => {
       if (o.option.indexOf('<origin>') > -1) {
@@ -407,29 +379,29 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
 
   it('doesn\'t fail if the parent doesn\'t define options', () => {
     sinon.stub(Command.prototype, 'options').callsFake(() => { return undefined; });
-    const options = (tenantCdnOriginRemoveCommand.options() as CommandOption[]);
+    const options = (command.options() as CommandOption[]);
     Utils.restore(Command.prototype.options);
     assert(options.length > 0);
   });
 
   it('accepts Public SharePoint Online CDN type', () => {
-    const actual = (tenantCdnOriginRemoveCommand.validate() as CommandValidate)({ options: { type: 'Public' } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: 'Public' } });
     assert(actual);
   });
 
   it('accepts Private SharePoint Online CDN type', () => {
-    const actual = (tenantCdnOriginRemoveCommand.validate() as CommandValidate)({ options: { type: 'Private' } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: 'Private' } });
     assert(actual);
   });
 
   it('rejects invalid SharePoint Online CDN type', () => {
     const type = 'foo';
-    const actual = (tenantCdnOriginRemoveCommand.validate() as CommandValidate)({ options: { type: type } });
+    const actual = (command.validate() as CommandValidate)({ options: { type: type } });
     assert.equal(actual, `${type} is not a valid CDN type. Allowed values are Public|Private`);
   });
 
   it('doesn\'t fail validation if the optional type option not specified', () => {
-    const actual = (tenantCdnOriginRemoveCommand.validate() as CommandValidate)({ options: {} });
+    const actual = (command.validate() as CommandValidate)({ options: {} });
     assert(actual);
   });
 
@@ -440,8 +412,8 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
       helpInformation: () => { }
     };
     const find = sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnOriginRemoveCommand.help() as CommandHelp)({}, helpLog);
-    assert(find.calledWith(commands.TENANT_CDN_ORIGIN_REMOVE));
+    (command.help() as CommandHelp)({}, helpLog);
+    assert(find.calledWith(commands.CDN_ORIGIN_SET));
   });
 
   it('has help with examples', () => {
@@ -451,7 +423,7 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
       helpInformation: () => { }
     };
     sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnOriginRemoveCommand.help() as CommandHelp)({}, log);
+    (command.help() as CommandHelp)({}, log);
     let containsExamples: boolean = false;
     _log.forEach(l => {
       if (l && l.indexOf('Examples:') > -1) {
@@ -468,7 +440,7 @@ describe(commands.TENANT_CDN_ORIGIN_REMOVE, () => {
     auth.site = new Site();
     auth.site.connected = true;
     auth.site.url = 'https://contoso-admin.sharepoint.com';
-    cmdInstance.action = tenantCdnOriginRemoveCommand.action();
+    cmdInstance.action = command.action();
     cmdInstance.action({ options: { verbose: true, confirm: true, key: 'existingproperty' }, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
       let containsError = false;
       log.forEach(l => {
