+++ src/o365/spo/commands/cdn-policy-set.ts
@@ -10,6 +10,7 @@ import {
   CommandValidate
 } from '../../../Command';
 import SpoCommand from '../SpoCommand';
+import Utils from '../../../Utils';
 
 const vorpal: Vorpal = require('../../../vorpal-init');
 
@@ -19,20 +20,23 @@ interface CommandArgs {
 
 interface Options extends VerboseOption {
   type: string;
+  policy: string;
+  value: string;
 }
 
-class SpoTenantCdnPolicyListCommand extends SpoCommand {
+class SpoCdnPolicySetCommand extends SpoCommand {
   public get name(): string {
-    return commands.TENANT_CDN_POLICY_LIST;
+    return commands.CDN_POLICY_SET;
   }
 
   public get description(): string {
-    return 'Lists CDN policies settings for the current SharePoint Online tenant';
+    return 'Sets CDN policy value for the current SharePoint Online tenant';
   }
 
   public getTelemetryProperties(args: CommandArgs): any {
     const telemetryProps: any = super.getTelemetryProperties(args);
     telemetryProps.cdnType = args.options.type || 'Public';
+    telemetryProps.policy = args.options.policy;
     return telemetryProps;
   }
 
@@ -57,7 +61,7 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
           cmd.log('');
         }
 
-        return this.getRequestDigest(cmd, this.verbose);
+        return this.getRequestDigest(cmd, this.verbose)
       })
       .then((res: ContextInfo): Promise<string> => {
         if (this.verbose) {
@@ -66,15 +70,25 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
           cmd.log('');
         }
 
-        cmd.log(`Retrieving configured policies for ${(cdnType === 1 ? 'Private' : 'Public')} CDN...`);
+        cmd.log(`Configuring policy on the ${(cdnType === 1 ? 'Private' : 'Public')} CDN. Please wait, this might take a moment...`);
+
+        let policyId: number = -1;
+        switch (args.options.policy) {
+          case "IncludeFileExtensions":
+            policyId = 0;
+            break;
+          case "ExcludeRestrictedSiteClassifications":
+            policyId = 1;
+            break;
+        }
 
         const requestOptions: any = {
           url: `${auth.site.url}/_vti_bin/client.svc/ProcessQuery`,
           headers: {
-            authorization: `Bearer ${auth.service.accessToken}`,
+            authorization: `Bearer ${auth.site.accessToken}`,
             'X-RequestDigest': res.FormDigestValue
           },
-          body: `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="GetTenantCdnPolicies" Id="7" ObjectPathId="3"><Parameters><Parameter Type="Enum">${cdnType}</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="3" Name="${auth.site.tenantId}" /></ObjectPaths></Request>`
+          body: `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="SetTenantCdnPolicy" Id="12" ObjectPathId="8"><Parameters><Parameter Type="Enum">${cdnType}</Parameter><Parameter Type="Enum">${policyId}</Parameter><Parameter Type="String">${Utils.escapeXml(args.options.value)}</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="8" Name="${auth.site.tenantId}" /></ObjectPaths></Request>`
         };
 
         if (this.verbose) {
@@ -98,12 +112,7 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
           cmd.log(vorpal.chalk.red(`Error: ${response.ErrorInfo.ErrorMessage}`));
         }
         else {
-          const result: string[] = json[json.length - 1];
-          cmd.log('Configured policies:');
-          result.forEach(o => {
-            const kv: string[] = o.split(';');
-            cmd.log(`${kv[0]}: ${kv[1]}`);
-          });
+          cmd.log(vorpal.chalk.green('DONE'));
         }
         cb();
       }, (err: any): void => {
@@ -113,11 +122,22 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
   }
 
   public options(): CommandOption[] {
-    const options: CommandOption[] = [{
-      option: '-t, --type [type]',
-      description: 'Type of CDN to manage. Public|Private. Default Public',
-      autocomplete: ['Public', 'Private']
-    }];
+    const options: CommandOption[] = [
+      {
+        option: '-t, --type [type]',
+        description: 'Type of CDN to manage. Public|Private. Default Public',
+        autocomplete: ['Public', 'Private']
+      },
+      {
+        option: '-p, --policy <policy>',
+        description: 'CDN policy to configure. IncludeFileExtensions|ExcludeRestrictedSiteClassifications',
+        autocomplete: ['IncludeFileExtensions', 'ExcludeRestrictedSiteClassifications']
+      },
+      {
+        option: '-v, --value <value>',
+        description: 'Value for the policy to configure'
+      }
+    ];
 
     const parentOptions: CommandOption[] | undefined = super.options();
     if (parentOptions) {
@@ -137,6 +157,12 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
         }
       }
 
+      if (!args.options.policy ||
+        (args.options.policy !== 'IncludeFileExtensions' &&
+        args.options.policy !== 'ExcludeRestrictedSiteClassifications')) {
+        return `${args.options.policy} is not a valid CDN policy. Allowed values are IncludeFileExtensions|ExcludeRestrictedSiteClassifications`;
+      }
+
       return true;
     };
   }
@@ -144,14 +170,14 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
   public help(): CommandHelp {
     return function (args: CommandArgs, log: (help: string) => void): void {
       const chalk = vorpal.chalk;
-      log(vorpal.find(commands.TENANT_CDN_POLICY_LIST).helpInformation());
+      log(vorpal.find(commands.CDN_POLICY_SET).helpInformation());
       log(
         `  ${chalk.yellow('Important:')} before using this command, connect to a SharePoint Online tenant admin site,
   using the ${chalk.blue(commands.CONNECT)} command.
         
   Remarks:
 
-    To list the policies of an Office 365 CDN, you have to first connect to a tenant admin site using the
+    To set the policy of an Office 365 CDN, you have to first connect to a tenant admin site using the
     ${chalk.blue(commands.CONNECT)} command, eg. ${chalk.grey(`${config.delimiter} ${commands.CONNECT} https://contoso-admin.sharepoint.com`)}.
     If you are connected to a different site and will try to manage tenant properties,
     you will get an error.
@@ -161,11 +187,8 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
 
   Examples:
   
-    ${chalk.grey(config.delimiter)} ${commands.TENANT_CDN_POLICY_LIST}
-      shows the list of policies configured for the Public CDN
-
-    ${chalk.grey(config.delimiter)} ${commands.TENANT_CDN_POLICY_LIST} -t Private
-      shows the list of policies configured for the Private CDN
+    ${chalk.grey(config.delimiter)} ${commands.CDN_POLICY_SET} -t Public -p IncludeFileExtensions -v CSS,EOT,GIF,ICO,JPEG,JPG,JS,MAP,PNG,SVG,TTF,WOFF,JSON
+      sets the list of extensions supported by the Public CDN
 
   More information:
 
@@ -176,4 +199,4 @@ class SpoTenantCdnPolicyListCommand extends SpoCommand {
   }
 }
 
-module.exports = new SpoTenantCdnPolicyListCommand();
+module.exports = new SpoCdnPolicySetCommand();
