--- src/o365/spo/commands/tenant-cdn-origin-set.spec.ts
@@ -1,461 +1,0 @@
-import commands from '../commands';
-import Command, { CommandHelp, CommandValidate, CommandOption } from '../../../Command';
-import * as sinon from 'sinon';
-import appInsights from '../../../appInsights';
-import auth, { Site } from '../SpoAuth';
-const tenantCdnOriginSetCommand: Command = require('./tenant-cdn-origin-set');
-import * as assert from 'assert';
-import * as request from 'request-promise-native';
-import config from '../../../config';
-import Utils from '../../../Utils';
-
-describe(commands.TENANT_CDN_ORIGIN_SET, () => {
-  let vorpal: Vorpal;
-  let log: string[];
-  let cmdInstance: any;
-  let trackEvent: any;
-  let telemetry: any;
-  let requests: any[];
-
-  before(() => {
-    sinon.stub(auth, 'ensureAccessToken').callsFake(() => { return Promise.resolve('ABC'); });
-    trackEvent = sinon.stub(appInsights, 'trackEvent').callsFake((t) => {
-      telemetry = t;
-    });
-    sinon.stub(request, 'post').callsFake((opts) => {
-      requests.push(opts);
-
-      if (opts.url.indexOf('/_api/contextinfo') > -1) {
-        if (opts.headers.authorization &&
-          opts.headers.authorization.indexOf('Bearer ') === 0 &&
-          opts.headers.accept &&
-          opts.headers.accept.indexOf('application/json') === 0) {
-          return Promise.resolve({ FormDigestValue: 'abc' });
-        }
-      }
-
-      if (opts.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1) {
-        if (opts.headers.authorization &&
-          opts.headers.authorization.indexOf('Bearer ') === 0 &&
-          opts.headers['X-RequestDigest'] &&
-          opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
-            return Promise.resolve(JSON.stringify([
-              {
-                "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": null, "TraceCorrelationId": "a05d299e-0036-4000-8546-cfc42dc07fd2"
-              }, 42, [
-                "*\u002fMASTERPAGE", "*\u002fSTYLE LIBRARY", "*\u002fCLIENTSIDEASSETS", "*\u002fCDN (configuration pending)"
-              ]
-            ]));
-          }
-        }
-      }
-
-      return Promise.reject('Invalid request');
-    });
-  });
-
-  beforeEach(() => {
-    vorpal = require('../../../vorpal-init');
-    log = [];
-    cmdInstance = {
-      log: (msg: string) => {
-        log.push(msg);
-      }
-    };
-    auth.site = new Site();
-    telemetry = null;
-    requests = [];
-  });
-
-  afterEach(() => {
-    Utils.restore(vorpal.find);
-  });
-
-  after(() => {
-    Utils.restore([
-      appInsights.trackEvent,
-      auth.ensureAccessToken,
-      request.post
-    ]);
-  });
-
-  it('has correct name', () => {
-    assert.equal(tenantCdnOriginSetCommand.name.startsWith(commands.TENANT_CDN_ORIGIN_SET), true);
-  });
-
-  it('has a description', () => {
-    assert.notEqual(tenantCdnOriginSetCommand.description, null);
-  });
-
-  it('calls telemetry', (done) => {
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
-      try {
-        assert(trackEvent.called);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('logs correct telemetry event', (done) => {
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: {}, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
-      try {
-        assert.equal(telemetry.name, commands.TENANT_CDN_ORIGIN_SET);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('aborts when not connected to a SharePoint site', (done) => {
-    auth.site = new Site();
-    auth.site.connected = false;
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
-      let returnsCorrectValue: boolean = false;
-      log.forEach(l => {
-        if (l && l.indexOf('Connect to a SharePoint Online site first') > -1) {
-          returnsCorrectValue = true;
-        }
-      });
-      try {
-        assert(returnsCorrectValue);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('aborts when not connected to a SharePoint tenant admin site', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso.sharepoint.com';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true }, appCatalogUrl: 'https://contoso.sharepoint.com/sites/appcatalog' }, () => {
-      let returnsCorrectValue: boolean = false;
-      log.forEach(l => {
-        if (l && l.indexOf(`${auth.site.url} is not a tenant admin site`) > -1) {
-          returnsCorrectValue = true;
-        }
-      });
-      try {
-        assert(returnsCorrectValue);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('sets CDN origin on the public CDN when Public type specified', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn', type: 'Public' } }, () => {
-      let setRequestIssued = false;
-      requests.forEach(r => {
-        if (r.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1 &&
-          r.headers.authorization &&
-          r.headers.authorization.indexOf('Bearer ') === 0 &&
-          r.headers['X-RequestDigest'] &&
-          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
-          setRequestIssued = true;
-        }
-      });
-
-      try {
-        assert(setRequestIssued);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('sets CDN origin on the private CDN when Private type specified', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn', type: 'Private' } }, () => {
-      let setRequestIssued = false;
-      requests.forEach(r => {
-        if (r.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1 &&
-          r.headers.authorization &&
-          r.headers.authorization.indexOf('Bearer ') === 0 &&
-          r.headers['X-RequestDigest'] &&
-          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">1</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
-          setRequestIssued = true;
-        }
-      });
-
-      try {
-        assert(setRequestIssued);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('sets CDN origin on the public CDN when no type specified', (done) => {
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: false, origin: '*/cdn' } }, () => {
-      let setRequestIssued = false;
-      requests.forEach(r => {
-        if (r.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1 &&
-          r.headers.authorization &&
-          r.headers.authorization.indexOf('Bearer ') === 0 &&
-          r.headers['X-RequestDigest'] &&
-          r.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
-          setRequestIssued = true;
-        }
-      });
-
-      try {
-        assert(setRequestIssued);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('correctly handles trying to set CDN origin that has already been set', (done) => {
-    Utils.restore(request.post);
-    sinon.stub(request, 'post').callsFake((opts) => {
-      requests.push(opts);
-
-      if (opts.url.indexOf('/_api/contextinfo') > -1) {
-        if (opts.headers.authorization &&
-          opts.headers.authorization.indexOf('Bearer ') === 0 &&
-          opts.headers.accept &&
-          opts.headers.accept.indexOf('application/json') === 0) {
-          return Promise.resolve({ FormDigestValue: 'abc' });
-        }
-      }
-
-      if (opts.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1) {
-        if (opts.headers.authorization &&
-          opts.headers.authorization.indexOf('Bearer ') === 0 &&
-          opts.headers['X-RequestDigest'] &&
-          opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">*/cdn</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
-            return Promise.resolve(JSON.stringify([
-              {
-                "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": {
-                  "ErrorMessage": "The library is already registered as a CDN origin.", "ErrorValue": null, "TraceCorrelationId": "965d299e-a0c6-4000-8546-cc244881a129", "ErrorCode": -1, "ErrorTypeName": "Microsoft.SharePoint.PublicCdn.TenantCdnAdministrationException"
-                }, "TraceCorrelationId": "965d299e-a0c6-4000-8546-cc244881a129"
-              }
-            ]));
-          }
-        }
-      }
-
-      return Promise.reject('Invalid request');
-    });
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true, origin: '*/cdn', type: 'Public' } }, () => {
-      let correctErrorReported = false;
-      log.forEach(l => {
-        if (l && typeof l === 'string' &&
-          l.indexOf('The library is already registered as a CDN origin.') > -1) {
-          correctErrorReported = true;
-        }
-      });
-
-      try {
-        assert(correctErrorReported);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('escapes XML in user input', (done) => {
-    Utils.restore(request.post);
-    sinon.stub(request, 'post').callsFake((opts) => {
-      requests.push(opts);
-
-      if (opts.url.indexOf('/_api/contextinfo') > -1) {
-        if (opts.headers.authorization &&
-          opts.headers.authorization.indexOf('Bearer ') === 0 &&
-          opts.headers.accept &&
-          opts.headers.accept.indexOf('application/json') === 0) {
-          return Promise.resolve({ FormDigestValue: 'abc' });
-        }
-      }
-
-      if (opts.url.indexOf('/_vti_bin/client.svc/ProcessQuery') > -1) {
-        if (opts.headers.authorization &&
-          opts.headers.authorization.indexOf('Bearer ') === 0 &&
-          opts.headers['X-RequestDigest'] &&
-          opts.body) {
-          if (opts.body === `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><Method Name="AddTenantCdnOrigin" Id="27" ObjectPathId="23"><Parameters><Parameter Type="Enum">0</Parameter><Parameter Type="String">&lt;*/CDN&gt;</Parameter></Parameters></Method></Actions><ObjectPaths><Identity Id="23" Name="abc" /></ObjectPaths></Request>`) {
-            return Promise.resolve(JSON.stringify([
-              {
-                "SchemaVersion": "15.0.0.0", "LibraryVersion": "16.0.7018.1204", "ErrorInfo": null, "TraceCorrelationId": "a05d299e-0036-4000-8546-cfc42dc07fd2"
-              }, 42, [
-                "*\u002fMASTERPAGE", "*\u002fSTYLE LIBRARY", "*\u002fCLIENTSIDEASSETS", "*\u002fCDN (configuration pending)"
-              ]
-            ]));
-          }
-        }
-      }
-
-      return Promise.reject('Invalid request');
-    });
-
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    auth.site.tenantId = 'abc';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true, origin: '<*/CDN>' } }, () => {
-      let isDone = false;
-      log.forEach(l => {
-        if (l && typeof l === 'string' && l.indexOf('DONE')) {
-          isDone = true;
-        }
-      });
-
-      try {
-        assert(isDone);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-
-  it('supports verbose mode', () => {
-    const options = (tenantCdnOriginSetCommand.options() as CommandOption[]);
-    let containsVerboseOption = false;
-    options.forEach(o => {
-      if (o.option === '--verbose') {
-        containsVerboseOption = true;
-      }
-    });
-    assert(containsVerboseOption);
-  });
-
-  it('requires CDN origin name', () => {
-    const options = (tenantCdnOriginSetCommand.options() as CommandOption[]);
-    let requiresCdnOriginName = false;
-    options.forEach(o => {
-      if (o.option.indexOf('<origin>') > -1) {
-        requiresCdnOriginName = true;
-      }
-    });
-    assert(requiresCdnOriginName);
-  });
-
-  it('doesn\'t fail if the parent doesn\'t define options', () => {
-    sinon.stub(Command.prototype, 'options').callsFake(() => { return undefined; });
-    const options = (tenantCdnOriginSetCommand.options() as CommandOption[]);
-    Utils.restore(Command.prototype.options);
-    assert(options.length > 0);
-  });
-
-  it('accepts Public SharePoint Online CDN type', () => {
-    const actual = (tenantCdnOriginSetCommand.validate() as CommandValidate)({ options: { type: 'Public' } });
-    assert(actual);
-  });
-
-  it('accepts Private SharePoint Online CDN type', () => {
-    const actual = (tenantCdnOriginSetCommand.validate() as CommandValidate)({ options: { type: 'Private' } });
-    assert(actual);
-  });
-
-  it('rejects invalid SharePoint Online CDN type', () => {
-    const type = 'foo';
-    const actual = (tenantCdnOriginSetCommand.validate() as CommandValidate)({ options: { type: type } });
-    assert.equal(actual, `${type} is not a valid CDN type. Allowed values are Public|Private`);
-  });
-
-  it('doesn\'t fail validation if the optional type option not specified', () => {
-    const actual = (tenantCdnOriginSetCommand.validate() as CommandValidate)({ options: {} });
-    assert(actual);
-  });
-
-  it('has help referring to the right command', () => {
-    const _helpLog: string[] = [];
-    const helpLog = (msg: string) => { _helpLog.push(msg); }
-    const cmd: any = {
-      helpInformation: () => { }
-    };
-    const find = sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnOriginSetCommand.help() as CommandHelp)({}, helpLog);
-    assert(find.calledWith(commands.TENANT_CDN_ORIGIN_SET));
-  });
-
-  it('has help with examples', () => {
-    const _log: string[] = [];
-    const log = (msg: string) => { _log.push(msg); }
-    const cmd: any = {
-      helpInformation: () => { }
-    };
-    sinon.stub(vorpal, 'find').callsFake(() => cmd);
-    (tenantCdnOriginSetCommand.help() as CommandHelp)({}, log);
-    let containsExamples: boolean = false;
-    _log.forEach(l => {
-      if (l && l.indexOf('Examples:') > -1) {
-        containsExamples = true;
-      }
-    });
-    Utils.restore(vorpal.find);
-    assert(containsExamples);
-  });
-
-  it('correctly handles lack of valid access token', (done) => {
-    Utils.restore(auth.ensureAccessToken);
-    sinon.stub(auth, 'ensureAccessToken').callsFake(() => { return Promise.reject(new Error('Error getting access token')); });
-    auth.site = new Site();
-    auth.site.connected = true;
-    auth.site.url = 'https://contoso-admin.sharepoint.com';
-    cmdInstance.action = tenantCdnOriginSetCommand.action();
-    cmdInstance.action({ options: { verbose: true, confirm: true, key: 'existingproperty' }, appCatalogUrl: 'https://contoso-admin.sharepoint.com' }, () => {
-      let containsError = false;
-      log.forEach(l => {
-        if (typeof l === 'string' &&
-          l.indexOf('Error getting access token') > -1) {
-          containsError = true;
-        }
-      });
-      try {
-        assert(containsError);
-        done();
-      }
-      catch (e) {
-        done(e);
-      }
-    });
-  });
-});
